warning removed

+++ include/lattice/poly.h
@@ -6,13 +6,13 @@
 #include <math.h>  
 #include "lweparam.h"
 typedef  struct{
-	int n;
+	size_t n;
 	mpi **a;
 	mpi *q;
 }Poly_q;
 
 typedef  struct{
-	int n;
+	size_t n;
 	int *a;
 }Poly_2;
 
@@ -21,8 +21,8 @@ int polyWriteBuffer(Poly_q* f, void* buffer);
 int polyReadBuffer(Poly_q* f, const void* buffer);
 int poly2WriteBuffer(Poly_2* f, void* buffer);
 int poly2ReadBuffer(Poly_2* f, const void* buffer);
-void RandomPoly(Poly_q * f, int n , mpi* q, float deviation, int hash, int (*f_rng)(void *, unsigned char *, size_t), void *p_rng ); //-1 for unbiased rand
-void ZeroPoly(Poly_q * f, int n ,mpi* q);
+void RandomPoly(Poly_q * f,  size_t n , mpi* q, float deviation, int hash, int (*f_rng)(void *, unsigned char *, size_t), void *p_rng ); //-1 for unbiased rand
+void ZeroPoly(Poly_q * f,  size_t n ,mpi* q);
 void ZeroPoly_2(Poly_2 * f, int n);
 void polyMul(Poly_q * c, Poly_q * a, Poly_q * b);
 void polyMulConst(Poly_q * c, int a, Poly_q * b);
